Refactor folder structure and unify player extensions
