+++ Sources/PDVideoPlayer/Common/Effect/RippleEffectStore.swift
