+++ Sources/PDVideoPlayer/iOS/PDVideoPlayer.swift
@@ -107,60 +107,4 @@ public extension PDVideoPlayer where MenuContent == EmptyView {
     }
 }
 
-public extension PDVideoPlayer {
-    /// Sets a binding for the muted state.
-    func isMuted(_ binding: Binding<Bool>) -> Self {
-        var copy = self
-        copy.isMuted = binding
-        return copy
-    }
-    
-    /// Sets a binding for control visibility.
-    func controlsVisible(_ binding: Binding<Bool>) -> Self {
-        var copy = self
-        copy.controlsVisible = binding
-        return copy
-    }
-    
-    
-    /// Sets a close action for the player.
-    func closeAction(_ action: VideoPlayerCloseAction) -> Self {
-        var copy = self
-        copy.closeAction = action
-        return copy
-    }
-    
-    /// Convenience overload to set a close action using a simple closure.
-    func closeAction(_ action: @escaping (CGFloat) -> Void) -> Self {
-        var copy = self
-        copy.closeAction = VideoPlayerCloseAction(action)
-        return copy
-    }
-    
-    /// Sets an action triggered when long‑press state changes.
-    func longpressAction(_ action: VideoPlayerLongpressAction) -> Self {
-        var copy = self
-        copy.longpressAction = action
-        return copy
-    }
-    
-    /// Convenience overload to set a long‑press action using a simple closure.
-    func longpressAction(_ action: @escaping (Bool) -> Void) -> Self {
-        var copy = self
-        copy.longpressAction = VideoPlayerLongpressAction(action)
-        return copy
-    }
-    /// Sets the foreground color for player controls.
-    func playerForegroundColor(_ color: Color) -> Self {
-        var copy = self
-        copy.foregroundColor = color
-        return copy
-    }
-    /// Sets the pan gesture style used for dismissing the video.
-    func panGesture(_ gesture: PDVideoPlayerPanGesture) -> Self {
-        var copy = self
-        copy.panGesture = gesture
-        return copy
-    }
-}
 #endif
